+++ modules/CPDoS.py
@@ -90,7 +90,7 @@ def run_cpdos_modules(
         print(" ! Canceled by keyboard interrupt (Ctrl-C)")
         sys.exit()
     except Exception as e:
-        print(e)
+        #print(e)
         logger.exception(e)